Rename: src 디렉터리 생성 및 프로젝트 패키지화
