--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "@mediapipe/hands": "^0.4.1675469240"
-  }
-}
